requirements.txt
configparser
requests
tqdm
paramiko
wikipedia
langdetect
gtts
deep_translator
msspeech
wxpython
yt_dlp
youtube-search-python
mpv
httpx==0.27.0
beautifulsoup4
patool